backend/ml/requirements.txt
flask
torch
torchvision
transformers
Pillow